# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Skorvia API - Brand credibility and ad-readiness analyzer
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /subscriptions/plans:
    get:
      operationId: listPlans
      tags: [subscriptions]
      summary: List available subscription plans with localized pricing
      parameters:
        - name: currency
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Subscription plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SubscriptionPlan"
...
  /subscriptions/initiate-payment:
    post:
      operationId: initiatePayment
      tags: [subscriptions]
      summary: Initiate a Flutterwave payment for subscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InitiatePaymentBody"
      responses:
        "200":
          description: Payment initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentInitResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /subscriptions/verify-payment:
    post:
      operationId: verifyPayment
      tags: [subscriptions]
      summary: Verify a Flutterwave payment after redirect
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyPaymentBody"
      responses:
        "200":
          description: Payment verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentVerifyResponse"
        "400":
          description: Verification failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    SubscriptionPlan:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          type: number
        currency:
          type: string
        currencySymbol:
          type: string
        interval:
          type: string
          enum: [monthly, yearly]
        features:
          type: array
          items:
            type: string
        isPopular:
          type: boolean
      required:
        - id
        - name
        - description
        - price
        - currency
        - currencySymbol
        - interval
        - features
        - isPopular
...
    InitiatePaymentBody:
      type: object
      properties:
        planId:
          type: string
        email:
          type: string
        name:
          type: string
        currency:
          type: string
        redirectUrl:
          type: string
      required:
        - planId
        - email
        - name
        - currency
        - redirectUrl
    PaymentInitResponse:
      type: object
      properties:
        paymentLink:
          type: string
        txRef:
          type: string
      required:
        - paymentLink
        - txRef
    VerifyPaymentBody:
      type: object
      properties:
        transactionId:
          type: string
        txRef:
          type: string
      required:
        - transactionId
        - txRef
    PaymentVerifyResponse:
      type: object
      properties:
        success:
          type: boolean
        planId:
          type: string
        subscriptionId:
          type: integer
        message:
          type: string
      required:
        - success
        - message
...